# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This implements the Airport API Reference.",
        "title": "Airport API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/luggageLabel/{number}": {
            "get": {
                "description": "Get the luggage label for a specific ticket by number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lugagges"
                ],
                "summary": "Retrieve a ticket's luggage label",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved luggage label",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Lugagge"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/tickets": {
            "get": {
                "description": "get all tickets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get all tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Ticket"
                            }
                        }
                    },
                    "404": {
                        "description": "No tickets found",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Add a new ticket with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Add a new ticket",
                "parameters": [
                    {
                        "description": "Add a new ticket",
                        "name": "AddTicketBodyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_tickets.AddTicketBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created a new ticket",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {}
                    }
                }
            }
        },
        "/tickets/cpf/{cpf}": {
            "get": {
                "description": "get tickets by passenger's CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get tickets by CPF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Ticket"
                            }
                        }
                    },
                    "404": {
                        "description": "No tickets found",
                        "schema": {}
                    }
                }
            }
        },
        "/tickets/{id}": {
            "put": {
                "description": "update ticket by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ticket details to be updated",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_tickets.UpdateTicketBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted a ticket",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/voucher/{number}": {
            "get": {
                "description": "Get the voucher details for a specific ticket by number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vouchers"
                ],
                "summary": "Retrieve a ticket voucher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved voucher",
                        "schema": {
                            "$ref": "#/definitions/handlers_tickets.Voucher"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_krymancer_airport-api-go_pkg_common_models.Lugagge": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "ticket_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_krymancer_airport-api-go_pkg_common_models.Ticket": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "flight": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lugagge": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Lugagge"
                    }
                },
                "number": {
                    "type": "integer"
                },
                "passenger_id": {
                    "type": "integer"
                },
                "seat_number": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "integer"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "handlers_tickets.AddTicketBodyRequest": {
            "type": "object",
            "properties": {
                "flight": {
                    "type": "integer",
                    "format": "integer",
                    "example": 123
                },
                "lugagge": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Lugagge"
                    }
                },
                "number": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1
                },
                "passenger": {
                    "type": "integer",
                    "format": "integer",
                    "example": 456
                },
                "seat_number": {
                    "type": "string",
                    "format": "string",
                    "example": "A1"
                }
            }
        },
        "handlers_tickets.UpdateTicketBodyRequest": {
            "type": "object",
            "properties": {
                "flight": {
                    "type": "integer",
                    "format": "integer",
                    "example": 123
                },
                "lugagge": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Lugagge"
                    }
                },
                "number": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1
                },
                "passenger": {
                    "type": "integer",
                    "format": "integer",
                    "example": 456
                },
                "seat_number": {
                    "type": "string",
                    "format": "string",
                    "example": "A1"
                }
            }
        },
        "handlers_tickets.Voucher": {
            "type": "object",
            "properties": {
                "destination": {
                    "type": "string",
                    "format": "string",
                    "example": "Los Angeles"
                },
                "flight_number": {
                    "type": "integer",
                    "format": "integer",
                    "example": 5678
                },
                "luggage_checked": {
                    "type": "boolean",
                    "example": true
                },
                "number": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1234
                },
                "origin": {
                    "type": "string",
                    "format": "string",
                    "example": "New York"
                },
                "passenger_cpf": {
                    "type": "string",
                    "format": "string",
                    "example": "123.456.789-00"
                },
                "passenger_name": {
                    "type": "string",
                    "format": "string",
                    "example": "John Doe"
                },
                "seat_number": {
                    "type": "string",
                    "format": "string",
                    "example": "A1"
                }
            }
        }
    }
}